requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
asyncio
lxml>=4.9.0
jupyter>=1.0.0
PyYAML>=6.0
langgraph>=0.0.40
langchain-core>=0.1.0
